feat(extractor): extract generic instances (#89)
closes #87

For example:

```ts
export interface Foo<T> {
  bar: Foo<1>
}
```

`Foo<T>` will be indexed as `Foo` and its property `bar` will be an inline type of `Foo<1>` whose target (the generic) is `Foo`.

Note this does not extract what the type arguments of the generic instance to the generic are. In the above example the parameter is `T` and the argument is `1`.