feat(extractor): type arguments of generic instances (#95)
closes #94
closes #93

Now we extract the type arguments of generic instances. For example in
the following type "B" is an instance of generic "A". "B" has type argument
"1" while A has type parameter "T".

```ts
type A<T> = {}
type B = A<1>
```

We also now handle standard library generics. For example in the
following type "A" is a generic instance of standard library type "Record"
and its type arguments are "string" and "string".

```ts
type A = Record<string, string>
```

This also fixes a double-alias entry for union alias members.
This improves things for `got`/`nexus`/`graphql-request` extraction such that we're back out of explicit errors :) but back into infinite loops :(.